requirements.txt
argon2-cffi==18.1.0
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.1.4
cryptoshop==2.0.1
idna==2.6
Naked==0.1.31
pycparser==2.18
pycrypto==2.6.1
PyYAML==3.12
requests==2.18.4
shellescape==3.4.1
six==1.11.0
tqdm==4.19.6
urllib3==1.22
